Carro semi desenhado com malhas de poligonos

+++ Frogger/LightSource.cpp
@@ -25,20 +25,20 @@ void LightSource::setAmbient(GLdouble x, GLdouble y, GLdouble z, GLdouble w){ _a
 void LightSource::setDiffuse(GLdouble x, GLdouble y, GLdouble z, GLdouble w){ _diffuse.set(x, y, z, w); }
 void LightSource::setSpecular(GLdouble x, GLdouble y, GLdouble z, GLdouble w){ _specular.set(x,y,z,w); }
 void LightSource::draw(){
-	GLfloat light_position[] = { _position.getX(), _position.getY(), _position.getZ(), _position.getW() };
+	GLfloat light_position[] = { (GLfloat)_position.getX(), (GLfloat)_position.getY(), (GLfloat)_position.getZ(), (GLfloat)_position.getW() };
 	glLightfv(GL_LIGHT0 + _num, GL_POSITION, light_position);
 
-	GLfloat ambient[] = { _ambient.getX(), _ambient.getY(), _ambient.getZ(), _ambient.getW() };
+	GLfloat ambient[] = { (GLfloat)_ambient.getX(), (GLfloat)_ambient.getY(), (GLfloat)_ambient.getZ(), (GLfloat)_ambient.getW() };
 	glLightfv(GL_LIGHT0 + _num, GL_AMBIENT, ambient);
 
-	GLfloat diffuse[] = { _diffuse.getX(), _diffuse.getY(), _diffuse.getZ(), _diffuse.getW() };
+	GLfloat diffuse[] = { (GLfloat)_diffuse.getX(), (GLfloat)_diffuse.getY(), (GLfloat)_diffuse.getZ(), (GLfloat)_diffuse.getW() };
 	glLightfv(GL_LIGHT0 + _num, GL_DIFFUSE, diffuse);
 
-	GLfloat specular[] = { _specular.getX(), _specular.getY(), _specular.getZ(), _specular.getW() };
+	GLfloat specular[] = { (GLfloat)_specular.getX(), (GLfloat)_specular.getY(), (GLfloat)_specular.getZ(),(GLfloat) _specular.getW() };
 	glLightfv(GL_LIGHT0 + _num, GL_SPECULAR, specular);
 	glLightfv(GL_LIGHT0 + _num, GL_EXP, &_exponent);
 	if ( _direction.getX() != 0 || _direction.getY() != 0 || _direction.getZ() != 0){
-		GLfloat direction[] = { _direction.getX(), _direction.getY(), _direction.getZ() };
+		GLfloat direction[] = { (GLfloat)_direction.getX(),(GLfloat) _direction.getY(),(GLfloat) _direction.getZ() };
 		glLightfv(GL_LIGHT0 + _num, GL_SPOT_DIRECTION, direction);
 		glLightf(GL_LIGHT0 + _num, GL_SPOT_CUTOFF, 60);
 		/*GLfloat a = 0.009, b = 0.0015, c = 0.001;
